requirements.txt
amqp==5.2.0
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
astroid==2.15.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
bcrypt==4.0.1
beanie==1.22.6
beautifulsoup4==4.11.1
billiard==4.2.0
black==23.9.1
blessed==1.20.0
bpython==0.24
cachetools==5.3.3
celery==5.3.6
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.2
curtsies==0.4.2
cwcwidth==0.1.8
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
distlib==0.3.7
dnspython==2.4.2
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
entrypoints==0.4
executing==1.2.0
fastapi==0.103.1
filelock==3.12.2
google-api-core==2.17.1
google-auth==2.28.1
googleapis-common-protos==1.62.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.8.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==5.1.4
keyboard==0.13.5
kombu==5.3.7
lazy-model==0.2.0
lazy-object-proxy==1.9.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mypy==1.8.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.1
oauthlib==3.2.2
orjson==3.8.7
packaging==23.0
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.13
platformdirs==3.10.0
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==4.25.3
psutil==5.9.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.14.0
pylint==2.17.0
pylint-plugin-utils==0.7
pylint-pydantic==0.1.8
pymongo==4.5.0
pyobjc==10.1
pyobjc-core==10.1
pyobjc-framework-Accessibility==10.1
pyobjc-framework-Accounts==10.1
pyobjc-framework-AddressBook==10.1
pyobjc-framework-AdServices==10.1
pyobjc-framework-AdSupport==10.1
pyobjc-framework-AppleScriptKit==10.1
pyobjc-framework-AppleScriptObjC==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-AppTrackingTransparency==10.1
pyobjc-framework-AudioVideoBridging==10.1
pyobjc-framework-AuthenticationServices==10.1
pyobjc-framework-AutomaticAssessmentConfiguration==10.1
pyobjc-framework-Automator==10.1
pyobjc-framework-AVFoundation==10.1
pyobjc-framework-AVKit==10.1
pyobjc-framework-AVRouting==10.1
pyobjc-framework-BackgroundAssets==10.1
pyobjc-framework-BusinessChat==10.1
pyobjc-framework-CalendarStore==10.1
pyobjc-framework-CallKit==10.1
pyobjc-framework-CFNetwork==10.1
pyobjc-framework-Cinematic==10.1
pyobjc-framework-ClassKit==10.1
pyobjc-framework-CloudKit==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Collaboration==10.1
pyobjc-framework-ColorSync==10.1
pyobjc-framework-Contacts==10.1
pyobjc-framework-ContactsUI==10.1
pyobjc-framework-CoreAudio==10.1
pyobjc-framework-CoreAudioKit==10.1
pyobjc-framework-CoreBluetooth==10.1
pyobjc-framework-CoreData==10.1
pyobjc-framework-CoreHaptics==10.1
pyobjc-framework-CoreLocation==10.1
pyobjc-framework-CoreMedia==10.1
pyobjc-framework-CoreMediaIO==10.1
pyobjc-framework-CoreMIDI==10.1
pyobjc-framework-CoreML==10.1
pyobjc-framework-CoreMotion==10.1
pyobjc-framework-CoreServices==10.1
pyobjc-framework-CoreSpotlight==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-CoreWLAN==10.1
pyobjc-framework-CryptoTokenKit==10.1
pyobjc-framework-DataDetection==10.1
pyobjc-framework-DeviceCheck==10.1
pyobjc-framework-DictionaryServices==10.1
pyobjc-framework-DiscRecording==10.1
pyobjc-framework-DiscRecordingUI==10.1
pyobjc-framework-DiskArbitration==10.1
pyobjc-framework-DVDPlayback==10.1
pyobjc-framework-EventKit==10.1
pyobjc-framework-ExceptionHandling==10.1
pyobjc-framework-ExecutionPolicy==10.1
pyobjc-framework-ExtensionKit==10.1
pyobjc-framework-ExternalAccessory==10.1
pyobjc-framework-FileProvider==10.1
pyobjc-framework-FileProviderUI==10.1
pyobjc-framework-FinderSync==10.1
pyobjc-framework-FSEvents==10.1
pyobjc-framework-GameCenter==10.1
pyobjc-framework-GameController==10.1
pyobjc-framework-GameKit==10.1
pyobjc-framework-GameplayKit==10.1
pyobjc-framework-HealthKit==10.1
pyobjc-framework-ImageCaptureCore==10.1
pyobjc-framework-InputMethodKit==10.1
pyobjc-framework-InstallerPlugins==10.1
pyobjc-framework-InstantMessage==10.1
pyobjc-framework-Intents==10.1
pyobjc-framework-IntentsUI==10.1
pyobjc-framework-IOBluetooth==10.1
pyobjc-framework-IOBluetoothUI==10.1
pyobjc-framework-IOSurface==10.1
pyobjc-framework-iTunesLibrary==10.1
pyobjc-framework-KernelManagement==10.1
pyobjc-framework-LatentSemanticMapping==10.1
pyobjc-framework-LaunchServices==10.1
pyobjc-framework-libdispatch==10.1
pyobjc-framework-libxpc==10.1
pyobjc-framework-LinkPresentation==10.1
pyobjc-framework-LocalAuthentication==10.1
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.1
pyobjc-framework-MailKit==10.1
pyobjc-framework-MapKit==10.1
pyobjc-framework-MediaAccessibility==10.1
pyobjc-framework-MediaLibrary==10.1
pyobjc-framework-MediaPlayer==10.1
pyobjc-framework-MediaToolbox==10.1
pyobjc-framework-Metal==10.1
pyobjc-framework-MetalFX==10.1
pyobjc-framework-MetalKit==10.1
pyobjc-framework-MetalPerformanceShaders==10.1
pyobjc-framework-MetalPerformanceShadersGraph==10.1
pyobjc-framework-MetricKit==10.1
pyobjc-framework-MLCompute==10.1
pyobjc-framework-ModelIO==10.1
pyobjc-framework-MultipeerConnectivity==10.1
pyobjc-framework-NaturalLanguage==10.1
pyobjc-framework-NetFS==10.1
pyobjc-framework-Network==10.1
pyobjc-framework-NetworkExtension==10.1
pyobjc-framework-NotificationCenter==10.1
pyobjc-framework-OpenDirectory==10.1
pyobjc-framework-OSAKit==10.1
pyobjc-framework-OSLog==10.1
pyobjc-framework-PassKit==10.1
pyobjc-framework-PencilKit==10.1
pyobjc-framework-PHASE==10.1
pyobjc-framework-Photos==10.1
pyobjc-framework-PhotosUI==10.1
pyobjc-framework-PreferencePanes==10.1
pyobjc-framework-PushKit==10.1
pyobjc-framework-Quartz==10.1
pyobjc-framework-QuickLookThumbnailing==10.1
pyobjc-framework-ReplayKit==10.1
pyobjc-framework-SafariServices==10.1
pyobjc-framework-SafetyKit==10.1
pyobjc-framework-SceneKit==10.1
pyobjc-framework-ScreenCaptureKit==10.1
pyobjc-framework-ScreenSaver==10.1
pyobjc-framework-ScreenTime==10.1
pyobjc-framework-ScriptingBridge==10.1
pyobjc-framework-SearchKit==10.1
pyobjc-framework-Security==10.1
pyobjc-framework-SecurityFoundation==10.1
pyobjc-framework-SecurityInterface==10.1
pyobjc-framework-SensitiveContentAnalysis==10.1
pyobjc-framework-ServiceManagement==10.1
pyobjc-framework-SharedWithYou==10.1
pyobjc-framework-SharedWithYouCore==10.1
pyobjc-framework-ShazamKit==10.1
pyobjc-framework-Social==10.1
pyobjc-framework-SoundAnalysis==10.1
pyobjc-framework-Speech==10.1
pyobjc-framework-SpriteKit==10.1
pyobjc-framework-StoreKit==10.1
pyobjc-framework-Symbols==10.1
pyobjc-framework-SyncServices==10.1
pyobjc-framework-SystemConfiguration==10.1
pyobjc-framework-SystemExtensions==10.1
pyobjc-framework-ThreadNetwork==10.1
pyobjc-framework-UniformTypeIdentifiers==10.1
pyobjc-framework-UserNotifications==10.1
pyobjc-framework-UserNotificationsUI==10.1
pyobjc-framework-VideoSubscriberAccount==10.1
pyobjc-framework-VideoToolbox==10.1
pyobjc-framework-Virtualization==10.1
pyobjc-framework-Vision==10.1
pyobjc-framework-WebKit==10.1
pyparsing==3.1.2
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pyxdg==0.28
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.5
rsa==4.9
ruff==0.1.13
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.27.0
toml==0.10.2
tomlkit==0.11.6
tornado==6.2
traitlets==5.8.1
typing_extensions==4.8.0
tzdata==2024.1
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.23.2
uvloop==0.17.0
vine==5.1.0
virtualenv==20.24.3
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
wrapt==1.15.0
yarg==0.1.9
